Remove enum type from OData V2 Despite implementing enum for V2 the first mention of enum in specification is only in later versions of OData. Thus, it should not be possible to parse enum using V2 parser.
