+++ pyodata/model/elements.py
@@ -9,7 +9,7 @@ from typing import Union
 from pyodata.config import Config
 from pyodata.exceptions import PyODataModelError, PyODataException, PyODataParserError
 
-from pyodata.model.type_traits import TypTraits, EdmStructTypTraits, EnumTypTrait
+from pyodata.model.type_traits import TypTraits, EdmStructTypTraits
 
 
 IdentifierInfo = collections.namedtuple('IdentifierInfo', 'namespace name')
@@ -571,75 +571,6 @@ class StructType(Typ):
 
 class ComplexType(StructType):
     """Representation of Edm.ComplexType"""
-
-
-class EnumMember:
-    def __init__(self, parent, name, value):
-        self._parent = parent
-        self._name = name
-        self._value = value
-
-    def __str__(self):
-        return f"{self._parent.name}\'{self._name}\'"
-
-    @property
-    def name(self):
-        return self._name
-
-    @property
-    def value(self):
-        return self._value
-
-    @property
-    def parent(self):
-        return self._parent
-
-
-class EnumType(Identifier):
-    def __init__(self, name, is_flags, underlying_type, namespace):
-        super(EnumType, self).__init__(name)
-        self._member = list()
-        self._underlying_type = underlying_type
-        self._traits = TypTraits()
-        self._namespace = namespace
-
-        if is_flags == 'True':
-            self._is_flags = True
-        else:
-            self._is_flags = False
-
-    def __str__(self):
-        return f"{self.__class__.__name__}({self._name})"
-
-    def __getattr__(self, item):
-        member = next(filter(lambda x: x.name == item, self._member), None)
-        if member is None:
-            raise PyODataException(f'EnumType {self} has no member {item}')
-
-        return member
-
-    def __getitem__(self, item):
-        # If the item is type string then we want to check for members with that name instead
-        if isinstance(item, str):
-            return self.__getattr__(item)
-
-        member = next(filter(lambda x: x.value == int(item), self._member), None)
-        if member is None:
-            raise PyODataException(f'EnumType {self} has no member with value {item}')
-
-        return member
-
-    @property
-    def is_flags(self):
-        return self._is_flags
-
-    @property
-    def traits(self):
-        return EnumTypTrait(self)
-
-    @property
-    def namespace(self):
-        return self._namespace
 
 
 class EntityType(StructType):
