+++ pyodata/model/type_traits.py
@@ -178,21 +178,3 @@ class EdmStructTypTraits(TypTraits):
 
     def from_literal(self, value):
         return EdmStructTypeSerializer.from_json(self._edm_type, value)
-
-
-class EnumTypTrait(TypTraits):
-    def __init__(self, enum_type):
-        self._enum_type = enum_type
-
-    def to_literal(self, value):
-        return f'{value.parent.namespace}.{value}'
-
-    def from_json(self, value):
-        return getattr(self._enum_type, value)
-
-    def from_literal(self, value):
-        # remove namespaces
-        enum_value = value.split('.')[-1]
-        # remove enum type
-        name = enum_value.split("'")[1]
-        return getattr(self._enum_type, name)
