+++ pyodata/v4/build_functions.py
@@ -5,10 +5,9 @@ import itertools
 
 from pyodata.config import Config
 from pyodata.exceptions import PyODataParserError, PyODataModelError
-from pyodata.model.elements import ComplexType, Schema, EnumType, NullType, build_element, EntityType, Types,\
-    StructTypeProperty
+from pyodata.model.elements import ComplexType, Schema, NullType, build_element, EntityType, Types, StructTypeProperty
 from pyodata.policies import ParserError
-from pyodata.v4.elements import NavigationTypeProperty, NullProperty, ReferentialConstraint
+from pyodata.v4.elements import NavigationTypeProperty, NullProperty, ReferentialConstraint, EnumMember, EnumType
 
 
 # pylint: disable=protected-access,too-many-locals,too-many-branches,too-many-statements
@@ -132,3 +131,53 @@ def build_navigation_type_property(config: Config, node):
         partner,
         node.get('contains_target'),
         ref_cons)
+
+
+# pylint: disable=protected-access, too-many-locals
+def build_enum_type(config: Config, type_node, namespace):
+    ename = type_node.get('Name')
+    is_flags = type_node.get('IsFlags')
+
+    # namespace = kwargs['namespace']
+
+    underlying_type = type_node.get('UnderlyingType')
+
+    # https://docs.oasis-open.org/odata/odata-csdl-json/v4.01/csprd04/odata-csdl-json-v4.01-csprd04.html#sec_EnumerationType
+    if underlying_type is None:
+        underlying_type = 'Edm.Int32'
+
+    valid_types = {
+        'Edm.Byte': [0, 2 ** 8 - 1],
+        'Edm.Int16': [-2 ** 15, 2 ** 15 - 1],
+        'Edm.Int32': [-2 ** 31, 2 ** 31 - 1],
+        'Edm.Int64': [-2 ** 63, 2 ** 63 - 1],
+        'Edm.SByte': [-2 ** 7, 2 ** 7 - 1]
+    }
+
+    if underlying_type not in valid_types:
+        raise PyODataParserError(
+            f'Type {underlying_type} is not valid as underlying type for EnumType - must be one of {valid_types}')
+
+    mtype = Types.from_name(underlying_type, config)
+    etype = EnumType(ename, is_flags, mtype, namespace)
+
+    members = type_node.xpath('edm:Member', namespaces=config.namespaces)
+
+    next_value = 0
+    for member in members:
+        name = member.get('Name')
+        value = member.get('Value')
+
+        if value is not None:
+            next_value = int(value)
+
+        vtype = valid_types[underlying_type]
+        if not vtype[0] < next_value < vtype[1]:
+            raise PyODataParserError(f'Value {next_value} is out of range for type {underlying_type}')
+
+        emember = EnumMember(etype, name, next_value)
+        etype._member.append(emember)
+
+        next_value += 1
+
+    return etype
